requirements.txt
sqlalchemy
pyodbc
python-dotenv
psycopg2
pandas

